prepare 4.6.2 release (#155)

+++ src/main/java/com/launchdarkly/client/EventSummarizer.java
@@ -132,11 +132,11 @@ final class EventSummarizer {
   }
   
   static final class CounterValue {
-    int count;
+    long count;
     final JsonElement flagValue;
     final JsonElement defaultVal;
     
-    CounterValue(int count, JsonElement flagValue, JsonElement defaultVal) {
+    CounterValue(long count, JsonElement flagValue, JsonElement defaultVal) {
       this.count = count;
       this.flagValue = flagValue;
       this.defaultVal = defaultVal;
